move DanhMucController
